Add bug to check cpp action

+++ addNumbers.cpp
@@ -1,5 +1,5 @@
 #include "addNumbers.h"
 
-int add(int a, int b) {
+float add(int a, int b) {
     return a + b;
 }
